requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.4.1
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.13.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
evaluate==0.4.3
executing==2.2.0
filelock==3.17.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
google-play-scraper==1.2.7
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mysql-connector-python==9.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.28.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==309
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.5.3
Sastrawi==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
websocket-client==1.8.0
wordcloud==1.9.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
